web-tech-2/requirements.txt
uvicorn
fastapi
sqlalchemy
databases
asyncpg
psycopg2
bcrypt
python-jose
python-dotenv
jwt
pyjwt
pyJWT